feat: add persistent storage options to k8s driver
The k8s driver now supports the `persistent-volume-claim.requests.storage`
option. Setting this option changes the deployment into a statefulset
and creates a persistent volume claim where the buildkit data is stored.

+++ driver/kubernetes/driver.go
@@ -44,15 +44,17 @@ type Driver struct {
 
 	// if you add fields, remember to update docs:
 	// https://github.com/docker/docs/blob/main/content/build/drivers/kubernetes.md
-	minReplicas      int
-	deployment       *appsv1.Deployment
-	configMaps       []*corev1.ConfigMap
-	deploymentClient kubeclient.DeploymentClient
-	podClient        kubeclient.PodClient
-	configMapClient  kubeclient.ConfigMapClient
-	podChooser       podchooser.PodChooser
-	defaultLoad      bool
-	timeout          time.Duration
+	minReplicas       int
+	deployment        *appsv1.Deployment
+	statefulSet       *appsv1.StatefulSet
+	configMaps        []*corev1.ConfigMap
+	deploymentClient  kubeclient.DeploymentClient
+	statefulSetClient kubeclient.StatefulSetClient
+	podClient         kubeclient.PodClient
+	configMapClient   kubeclient.ConfigMapClient
+	podChooser        podchooser.PodChooser
+	defaultLoad       bool
+	timeout           time.Duration
 }
 
 func (d *Driver) IsMobyDriver() bool {
@@ -65,31 +67,18 @@ func (d *Driver) Config() driver.InitConfig {
 
 func (d *Driver) Bootstrap(ctx context.Context, l progress.Logger) error {
 	return progress.Wrap("[internal] booting buildkit", l, func(sub progress.SubLogger) error {
-		_, err := d.deploymentClient.Get(ctx, d.deployment.Name, metav1.GetOptions{})
-		if err != nil {
-			if !apierrors.IsNotFound(err) {
-				return errors.Wrapf(err, "error for bootstrap %q", d.deployment.Name)
-			}
-
-			for _, cfg := range d.configMaps {
-				// create ConfigMap first if exists
-				_, err = d.configMapClient.Create(ctx, cfg, metav1.CreateOptions{})
-				if err != nil {
-					if !apierrors.IsAlreadyExists(err) {
-						return errors.Wrapf(err, "error while calling configMapClient.Create for %q", cfg.Name)
-					}
-					_, err = d.configMapClient.Update(ctx, cfg, metav1.UpdateOptions{})
-					if err != nil {
-						return errors.Wrapf(err, "error while calling configMapClient.Update for %q", cfg.Name)
-					}
-				}
-			}
-
-			_, err = d.deploymentClient.Create(ctx, d.deployment, metav1.CreateOptions{})
-			if err != nil {
-				return errors.Wrapf(err, "error while calling deploymentClient.Create for %q", d.deployment.Name)
+		if d.deployment != nil {
+			if err := bootstrap(ctx, d, d.deploymentClient, d.deployment.Name, d.deployment); err != nil {
+				return err
 			}
 		}
+
+		if d.statefulSet != nil {
+			if err := bootstrap(ctx, d, d.statefulSetClient, d.statefulSet.Name, d.statefulSet); err != nil {
+				return err
+			}
+		}
+
 		return sub.Wrap(
 			fmt.Sprintf("waiting for %d pods to be ready, timeout: %s", d.minReplicas, units.HumanDuration(d.timeout)),
 			func() error {
@@ -98,11 +87,76 @@ func (d *Driver) Bootstrap(ctx context.Context, l progress.Logger) error {
 	})
 }
 
+type appClient[S any] interface {
+	Get(ctx context.Context, name string, opts metav1.GetOptions) (*S, error)
+	Create(ctx context.Context, spec *S, opts metav1.CreateOptions) (*S, error)
+	Delete(ctx context.Context, name string, opts metav1.DeleteOptions) error
+}
+
+func bootstrap[S any](ctx context.Context, d *Driver, client appClient[S], name string, spec *S) error {
+	if _, err := client.Get(ctx, name, metav1.GetOptions{}); err != nil {
+		if !apierrors.IsNotFound(err) {
+			return errors.Wrapf(err, "error for bootstrap %q", name)
+		}
+
+		for _, cfg := range d.configMaps {
+			// create ConfigMap first if exists
+			if _, err = d.configMapClient.Create(ctx, cfg, metav1.CreateOptions{}); err != nil {
+				if !apierrors.IsAlreadyExists(err) {
+					return errors.Wrapf(err, "error while calling configMapClient.Create for %q", cfg.Name)
+				}
+
+				if _, err = d.configMapClient.Update(ctx, cfg, metav1.UpdateOptions{}); err != nil {
+					return errors.Wrapf(err, "error while calling configMapClient.Update for %q", cfg.Name)
+				}
+			}
+		}
+
+		if _, err = client.Create(ctx, spec, metav1.CreateOptions{}); err != nil {
+			return errors.Wrapf(err, "error while calling Create for %q", d.deployment.Name)
+		}
+	}
+	return nil
+}
+
 func (d *Driver) wait(ctx context.Context) error {
+	if d.deployment != nil {
+		if err := d.waitDeployments(ctx); err != nil {
+			return err
+		}
+	}
+
+	if d.statefulSet != nil {
+		if err := d.waitStatefulSets(ctx); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func (d *Driver) waitDeployments(ctx context.Context) error {
+	return wait(ctx, d, d.deploymentClient, d.deployment.Name, func(s *appsv1.Deployment) error {
+		if s.Status.ReadyReplicas < int32(d.minReplicas) {
+			return errors.Errorf("expected %d replicas to be ready, got %d", d.minReplicas, s.Status.ReadyReplicas)
+		}
+		return nil
+	})
+}
+
+func (d *Driver) waitStatefulSets(ctx context.Context) error {
+	return wait(ctx, d, d.statefulSetClient, d.statefulSet.Name, func(s *appsv1.StatefulSet) error {
+		if s.Status.ReadyReplicas < int32(d.minReplicas) {
+			return errors.Errorf("expected %d replicas to be ready, got %d", d.minReplicas, s.Status.ReadyReplicas)
+		}
+		return nil
+	})
+}
+
+func wait[S any](ctx context.Context, d *Driver, client appClient[S], name string, check func(*S) error) error {
 	// TODO: use watch API
 	var (
 		err  error
-		depl *appsv1.Deployment
+		spec *S
 	)
 
 	timeoutChan := time.After(d.timeout)
@@ -116,31 +170,50 @@ func (d *Driver) wait(ctx context.Context) error {
 		case <-timeoutChan:
 			return err
 		case <-ticker.C:
-			depl, err = d.deploymentClient.Get(ctx, d.deployment.Name, metav1.GetOptions{})
+			spec, err = client.Get(ctx, name, metav1.GetOptions{})
 			if err == nil {
-				if depl.Status.ReadyReplicas >= int32(d.minReplicas) {
+				if err = check(spec); err == nil {
 					return nil
 				}
-				err = errors.Errorf("expected %d replicas to be ready, got %d", d.minReplicas, depl.Status.ReadyReplicas)
 			}
 		}
 	}
 }
 
-func (d *Driver) Info(ctx context.Context) (*driver.Info, error) {
-	depl, err := d.deploymentClient.Get(ctx, d.deployment.Name, metav1.GetOptions{})
-	if err != nil {
-		// TODO: return err if err != ErrNotFound
-		return &driver.Info{
-			Status: driver.Inactive,
-		}, nil
+func (d *Driver) Info(ctx context.Context) (_ *driver.Info, err error) {
+	var depl *appsv1.Deployment
+	if d.deployment != nil {
+		depl, err = d.deploymentClient.Get(ctx, d.deployment.Name, metav1.GetOptions{})
+		if err != nil {
+			// TODO: return err if err != ErrNotFound
+			return &driver.Info{
+				Status: driver.Inactive,
+			}, nil
+		}
+		if depl.Status.ReadyReplicas <= 0 {
+			return &driver.Info{
+				Status: driver.Stopped,
+			}, nil
+		}
 	}
-	if depl.Status.ReadyReplicas <= 0 {
-		return &driver.Info{
-			Status: driver.Stopped,
-		}, nil
+
+	var stat *appsv1.StatefulSet
+	if d.statefulSet != nil {
+		stat, err = d.statefulSetClient.Get(ctx, d.statefulSet.Name, metav1.GetOptions{})
+		if err != nil {
+			// TODO: return err if err != ErrNotFound
+			return &driver.Info{
+				Status: driver.Inactive,
+			}, nil
+		}
+		if depl.Status.ReadyReplicas <= 0 {
+			return &driver.Info{
+				Status: driver.Stopped,
+			}, nil
+		}
 	}
-	pods, err := podchooser.ListRunningPods(ctx, d.podClient, depl)
+
+	pods, err := podchooser.ListRunningPods(ctx, d.podClient, depl, stat)
 	if err != nil {
 		return nil, err
 	}
@@ -182,11 +255,22 @@ func (d *Driver) Rm(ctx context.Context, force, rmVolume, rmDaemon bool) error {
 		return nil
 	}
 
-	if err := d.deploymentClient.Delete(ctx, d.deployment.Name, metav1.DeleteOptions{}); err != nil {
-		if !apierrors.IsNotFound(err) {
-			return errors.Wrapf(err, "error while calling deploymentClient.Delete for %q", d.deployment.Name)
+	if d.deployment != nil {
+		if err := d.deploymentClient.Delete(ctx, d.deployment.Name, metav1.DeleteOptions{}); err != nil {
+			if !apierrors.IsNotFound(err) {
+				return errors.Wrapf(err, "error while calling deploymentClient.Delete for %q", d.deployment.Name)
+			}
 		}
 	}
+
+	if d.statefulSet != nil {
+		if err := d.statefulSetClient.Delete(ctx, d.statefulSet.Name, metav1.DeleteOptions{}); err != nil {
+			if !apierrors.IsNotFound(err) {
+				return errors.Wrapf(err, "error while calling statefulSetClient.Delete for %q", d.statefulSet.Name)
+			}
+		}
+	}
+
 	for _, cfg := range d.configMaps {
 		if err := d.configMapClient.Delete(ctx, cfg.Name, metav1.DeleteOptions{}); err != nil {
 			if !apierrors.IsNotFound(err) {
